Add floating ip, security group and ssh management.
Other changes:
- openstack actions and waiters functions were moved to dedicated
  class and added as a fixture
- iperf output parser was added
- variable for control cleanup objects was added
- improve logging

+++ tungsten_tests/fixtures/openstack_fixtures.py
@@ -1,12 +1,17 @@
+import hashlib
 import ipaddress
 import logging
+import os
+import paramiko
 import pytest
 
 import neutronclient.common.exceptions
 
 from tungsten_tests.clients.os_clients import OpenStackClientManager
 from tungsten_tests.helpers.common import download_file
-from tungsten_tests.settings import TFT_UBUNTU_IMG_URL, TFT_IMAGE_PATH
+from tungsten_tests.helpers.os_actions import OpenStackActions
+from tungsten_tests.settings import TFT_UBUNTU_IMG_URL, TFT_IMAGE_PATH, \
+    TFT_INSTANCE_KEYS_PATH, TFT_CLEANUP_SETUP
 
 logger = logging.getLogger()
 
@@ -22,28 +27,36 @@ def os_clients(config):
 
 
 @pytest.fixture(scope='session')
+def os_actions(config, os_clients, cleanup_session):
+    return OpenStackActions(os_clients, config, cleanup_session)
+
+
+@pytest.fixture(scope='session')
 def upload_images(config, os_clients):
     # Check if image is present
     for image in os_clients.glance.images.list():
         if image.name == config.os_ubuntu_img_name:
             config.os_ubuntu_img_id = image.id
-            logger.info("Image {} was found. Image id is {}".format(
+            logger.info("Image '{}' was found. Image ID: {}".format(
                 config.os_ubuntu_img_name, config.os_ubuntu_img_id))
     if not config.os_ubuntu_img_id:
-        logger.info("Image {} wasn't found.".format(config.os_ubuntu_img_name))
+        logger.info("Image '{}' wasn't found."
+                    "".format(config.os_ubuntu_img_name))
         file_path = download_file(TFT_UBUNTU_IMG_URL, TFT_IMAGE_PATH)
         image = os_clients.glance.images.create(name=config.os_ubuntu_img_name,
                                                 disk_format='qcow2',
                                                 container_format='bare',
                                                 visibility='public')
+        logger.info("Uploading image to glance...")
         os_clients.glance.images.upload(image.id, open(file_path, 'rb'))
         config.os_ubuntu_img_id = image.id
-        logger.info("Image {} is uploaded. Image id is {}".format(
+        logger.info("Image '{}' was uploaded. Image ID: {}".format(
             config.os_ubuntu_img_name, config.os_ubuntu_img_id))
     yield
     # Cleanup
-    logger.info("Delete image {}".format(config.os_ubuntu_img_id))
-    # os_clients.glance.images.delete(config.os_ubuntu_img_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete image {}".format(config.os_ubuntu_img_id))
+        os_clients.glance.images.delete(config.os_ubuntu_img_id)
 
 
 @pytest.fixture(scope='session')
@@ -52,22 +65,23 @@ def create_flavors(config, os_clients):
     for flavor in os_clients.nova.flavors.list():
         if flavor.name == config.os_ubuntu_flavor_name:
             config.os_ubuntu_flavor_id = flavor.id
-            logger.info("Flavor {} was found. Flavor id is {}".format(
+            logger.info("Flavor '{}' was found. Flavor ID: {}".format(
                 config.os_ubuntu_flavor_name, config.os_ubuntu_flavor_id))
     if not config.os_ubuntu_flavor_id:
-        logger.info("Flavor {} wasn't found."
+        logger.info("Flavor '{}' wasn't found."
                     "".format(config.os_ubuntu_flavor_name))
         flavor = os_clients.nova.flavors.create(
             name=config.os_ubuntu_flavor_name,
-            ram=512, vcpus=1, disk=1,
+            ram=512, vcpus=1, disk=3,
             is_public=True)
         config.os_ubuntu_flavor_id = flavor.id
-        logger.info("Flavor {} is created. Falvor id is {}".format(
+        logger.info("Flavor '{}' is created. Falvor ID: {}".format(
             config.os_ubuntu_flavor_name, config.os_ubuntu_flavor_id))
     yield
     # Cleanup
-    logger.info("Delete flavor {}".format(config.os_ubuntu_flavor_id))
-    os_clients.nova.flavors.delete(config.os_ubuntu_flavor_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete flavor {}".format(config.os_ubuntu_flavor_id))
+        os_clients.nova.flavors.delete(config.os_ubuntu_flavor_id)
 
 
 @pytest.fixture(scope='session')
@@ -77,22 +91,23 @@ def create_network(config, os_clients):
     for net in networks['networks']:
         if net['name'] == config.os_net_name:
             config.os_net_id = net['id']
-            logger.info("Network {} was found. Network id is {}".format(
+            logger.info("Network '{}' was found. Network ID: {}".format(
                 config.os_net_name, config.os_net_id))
     if not config.os_net_id:
-        logger.info("Network {} wasn't found.".format(config.os_net_name))
+        logger.info("Network '{}' wasn't found.".format(config.os_net_name))
         network = {
             'name': config.os_net_name,
             'admin_state_up': True
         }
         net = os_clients.neutron.create_network({'network': network})
         config.os_net_id = net['network']['id']
-        logger.info("Network {} is created. Network id is {}".format(
+        logger.info("Network '{}' is created. Network ID: {}".format(
             config.os_net_name, config.os_net_id))
     yield
     # Cleanup
-    logger.info("Delete network {}".format(config.os_net_id))
-    os_clients.neutron.delete_network(config.os_net_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete network {}".format(config.os_net_id))
+        os_clients.neutron.delete_network(config.os_net_id)
 
 
 @pytest.fixture(scope='session')
@@ -102,10 +117,10 @@ def create_subnet(config, os_clients, create_network):
     for subnet in subnets['subnets']:
         if subnet['name'] == config.os_subnet_name:
             config.os_subnet_id = subnet['id']
-            logger.info("Subnet {} was found. Subnet id is {}".format(
+            logger.info("Subnet '{}' was found. Subnet ID: {}".format(
                 config.os_subnet_name, config.os_subnet_id))
     if not config.os_subnet_id:
-        logger.info("Subnet {} wasn't found.".format(config.os_subnet_name))
+        logger.info("Subnet '{}' wasn't found.".format(config.os_subnet_name))
         ipv4_net = ipaddress.IPv4Network(unicode(config.os_subnet_cidr))
         ip_range = list(ipv4_net.hosts())
         subnet = {
@@ -118,12 +133,13 @@ def create_subnet(config, os_clients, create_network):
         }
         subnet = os_clients.neutron.create_subnet({'subnet': subnet})
         config.os_subnet_id = subnet['subnet']['id']
-        logger.info("Subnet {} is created. Subnet id is {}".format(
+        logger.info("Subnet '{}' is created. Subnet ID: {}".format(
             config.os_subnet_name, config.os_subnet_id))
     yield
     # Cleanup
-    logger.info("Delete subnet {}".format(config.os_subnet_id))
-    os_clients.neutron.delete_subnet(config.os_subnet_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete subnet {}".format(config.os_subnet_id))
+        os_clients.neutron.delete_subnet(config.os_subnet_id)
 
 
 @pytest.fixture(scope='session')
@@ -133,11 +149,11 @@ def create_sg(config, os_clients):
     for sg in sgs['security_groups']:
         if sg['name'] == config.os_sg_name:
             config.os_sg_id = sg['id']
-            logger.info("Security group {} was found. SG id is {}".format(
+            logger.info("Security group '{}' was found. SG ID: {}".format(
                 config.os_subnet_name, config.os_subnet_id))
             break
     if not config.os_sg_id:
-        logger.info("Security group {} wasn't found.".format(
+        logger.info("Security group '{}' wasn't found.".format(
             config.os_subnet_name))
         sg_body = {
             'name': config.os_sg_name
@@ -146,7 +162,7 @@ def create_sg(config, os_clients):
             {'security_group': sg_body}
         )
         config.os_sg_id = sg['security_group']['id']
-        logger.info("Security group {} is created. SG id is {}".format(
+        logger.info("Security group '{}' is created. SG ID: {}".format(
             config.os_subnet_name, config.os_subnet_id))
 
     # Add security group rules
@@ -199,8 +215,9 @@ def create_sg(config, os_clients):
             logger.warning("{}".format(e))
     yield
     # Cleanup
-    logger.info("Delete security group {}".format(config.os_sg_id))
-    os_clients.neutron.delete_security_group(config.os_sg_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete security group {}".format(config.os_sg_id))
+        os_clients.neutron.delete_security_group(config.os_sg_id)
 
 
 @pytest.fixture(scope='session')
@@ -209,16 +226,19 @@ def get_external_net(config, os_clients):
         net = os_clients.neutron.list_networks(name=config.os_ext_net_name)
         if len(net['networks']) == 1:
             if not net['networks'][0]['router:external']:
-                raise Exception("Network {} isn't external"
+                raise Exception("Network '{}' isn't external"
                                 "".format(config.os_ext_net_name))
             else:
                 config.os_ext_net_id = net['networks'][0]['id']
+                logger.info("External network '{}' was found. External network"
+                            " ID: {}".format(config.os_ext_net_name,
+                                             config.os_ext_net_id))
         elif len(net['networks']) > 1:
-            raise Exception("Can't determine external network {}. Was found "
+            raise Exception("Can't determine external network '{}'. Was found "
                             "several networks with such name: {}"
                             "".format(config.os_ext_net_name, net['networks']))
         else:
-            raise Exception("Can't find external network {}"
+            raise Exception("Can't find external network '{}'"
                             "".format(config.os_ext_net_name))
 
 
@@ -233,10 +253,10 @@ def create_router(config, os_clients, get_external_net):
                 raise Exception("Incorrect gateway network: {}. Should be: {}"
                                 "".format(ext_gw_net, config.os_ext_net_id))
             config.os_router_id = router['id']
-            logger.info("Router {} was found. Router id is {}".format(
+            logger.info("Router '{}' was found. Router ID: {}".format(
                 config.os_router_name, config.os_router_id))
     if not config.os_router_id:
-        logger.info("Router {} wasn't found.".format(config.os_router_name))
+        logger.info("Router '{}' wasn't found.".format(config.os_router_name))
         external_gateway_info = {
             'network_id': config.os_ext_net_id
         }
@@ -247,12 +267,13 @@ def create_router(config, os_clients, get_external_net):
         }
         router = os_clients.neutron.create_router({'router': router})
         config.os_router_id = router['router']['id']
-        logger.info("Router {} is created. Router id is {}".format(
+        logger.info("Router '{}' is created. Router ID: {}".format(
             config.os_router_name, config.os_router_id))
     yield
     # Cleanup
-    logger.info("Delete router {}".format(config.os_router_id))
-    os_clients.neutron.delete_router(config.os_router_id)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete router {}".format(config.os_router_id))
+        os_clients.neutron.delete_router(config.os_router_id)
 
 
 @pytest.fixture(scope='session')
@@ -266,10 +287,10 @@ def router_attach_subnet(config, os_clients, create_subnet, create_router):
                 raise Exception("Incorrect port network: {}. Should be: {}"
                                 "".format(net_id, config.os_net_id))
             config.os_port_id = port['id']
-            logger.info("Port {} was found. Port id is {}".format(
+            logger.info("Port '{}' was found. Port ID: {}".format(
                 config.os_port_name, config.os_port_id))
     if not config.os_port_id:
-        logger.info("Port {} wasn't found.".format(config.os_port_name))
+        logger.info("Port '{}' wasn't found.".format(config.os_port_name))
         interface = {
             "subnet_id": config.os_subnet_id
         }
@@ -280,10 +301,70 @@ def router_attach_subnet(config, os_clients, create_subnet, create_router):
             "name": config.os_port_name
         }
         os_clients.neutron.update_port(config.os_port_id, {'port': port})
-        logger.info("Port {}:{} was attached to router {}:{}".format(
-            config.os_port_name, config.os_port_id, config.os_router_name,
-            config.os_router_id))
+        logger.info("Port '{}' ID: {} was attached to router '{}' ID: {}"
+                    "".format(config.os_port_name, config.os_port_id,
+                              config.os_router_name, config.os_router_id))
     yield
     # Cleanup
-    logger.info("Detach port {}".format(config.os_port_id))
-    os_clients.neutron.remove_interface_router(config.os_router_id, interface)
+    if TFT_CLEANUP_SETUP:
+        logger.info("Detach port {}".format(config.os_port_id))
+        os_clients.neutron.remove_interface_router(config.os_router_id,
+                                                   interface)
+
+
+@pytest.fixture(scope='session')
+def create_keypair(config, os_clients):
+    # Check if keypair exist
+    keypairs = os_clients.nova.keypairs.list()
+    key_file = TFT_INSTANCE_KEYS_PATH + "/private_key"
+    for key in keypairs:
+        if key.name == config.os_keypair_name:
+            config.os_keypair_id = key.id
+            logger.info("Keypair '{}' was found."
+                        "".format(config.os_keypair_name))
+            break
+    if os.path.isfile(key_file) and config.os_keypair_id:
+        logger.info("Check ssh Fingerprints.\nPrivate key: {}"
+                    "".format(key_file))
+        os_private_key = paramiko.RSAKey.from_private_key_file(key_file)
+        sshFingerprint = hashlib.md5(os_private_key.__str__()).hexdigest()
+        printableFingerprint = ':'.join(
+            a + b for a, b in zip(sshFingerprint[::2], sshFingerprint[1::2]))
+        key = os_clients.nova.keypairs.get(config.os_keypair_id)
+        if printableFingerprint == key.fingerprint:
+            logger.info("The ssh Fingerprints match: {} == {}"
+                        "".format(printableFingerprint, key.fingerprint))
+            config.os_private_key = key_file
+        else:
+            logger.warning("The ssh Fingerprints don't match: {} != {}\n"
+                           "Delete keypair {}".format(printableFingerprint,
+                                                      key.fingerprint,
+                                                      key.id))
+            os_clients.nova.keypairs.delete(key.id)
+            config.os_keypair_id = None
+    elif not os.path.isfile(key_file) and config.os_keypair_id:
+        logger.warning("Private key not found: {}.\n"
+                       "Delete keypair {}".format(key_file, key.id))
+        os_clients.nova.keypairs.delete(key.id)
+        config.os_keypair_id = None
+
+    if not config.os_keypair_id:
+        logger.info("Keypair '{}' wasn't found."
+                    "".format(config.os_keypair_name))
+        keypair = os_clients.nova.keypairs.create(config.os_keypair_name)
+        config.os_keypair_id = keypair.id
+        config.os_private_key = keypair.private_key
+        with open(key_file, "w+") as f:
+            f.write(keypair.private_key)
+        os.chmod(key_file, 0o600)
+        config.os_private_key = key_file
+        logger.info("Keypair '{}' is created. Keypair ID: {}.\n"
+                    "Private key is stored in {}."
+                    "".format(config.os_keypair_name,
+                              config.os_keypair_id,
+                              key_file))
+    yield
+    # Cleanup
+    if TFT_CLEANUP_SETUP:
+        logger.info("Delete keypair {}".format(config.os_keypair_id))
+        os_clients.nova.keypairs.delete(config.os_keypair_id)
